implemented find closest in projectile, for this we dont need to sort for beam, we only sort collisions if we have more then one target in line homing projectile now uses faster find closest, only checks collision with closest entity homing projectile now less crazy

+++ src/level/space_level.h
@@ -23,7 +23,7 @@ public:
 
 private:
 
-	void UpdateStars(StarIter start, StarIter end, int index,  float worldSpeed, float dt) const;
+	void UpdateStars(StarIter start, StarIter end, size_t index, float worldSpeed, float dt) const;
 
 private:
 	std::shared_ptr<IThreadedWorkload> threadableWorkload;
